// typechain-types/ContentLedgerAnchor.ts
export interface ContentLedgerAnchorInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "admin"
      | "anchor"
      | "anchorBlocks"
      | "anchorHistory"
      | "anchorTimestamps"
      | "getAnchorCount"
      | "getBlockNumber"
      | "getTimestamp"
      | "isAnchored"
      | "totalAnchors"
      | "transferAdmin"
      | "verify"
  ): FunctionFragment;

  getEvent(
    nameOrSignatureOrTopic: "AnchorSubmitted" | "ContentVerified"
  ): EventFragment;

  encodeFunctionData(functionFragment: "admin", values?: undefined): string;
  encodeFunctionData(functionFragment: "anchor", values: [BytesLike]): string;
  encodeFunctionData(
    functionFragment: "anchorBlocks",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "anchorHistory",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "anchorTimestamps",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "getAnchorCount",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "getBlockNumber",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "getTimestamp",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "isAnchored",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "totalAnchors",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "transferAdmin",
    values: [AddressLike]
  ): string;
  encodeFunctionData(functionFragment: "verify", values: [BytesLike]): string;

  decodeFunctionResult(functionFragment: "admin", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "anchor", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "anchorBlocks",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "anchorHistory",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "anchorTimestamps",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getAnchorCount",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getBlockNumber",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getTimestamp",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "isAnchored", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "totalAnchors",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "transferAdmin",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "verify", data: BytesLike): Result;
}

export namespace AnchorSubmittedEvent {
  export type InputTuple = [
    merkleRoot: BytesLike,
    blockTimestamp: BigNumberish,
    blockNumber: BigNumberish
  ];
  export type OutputTuple = [
    merkleRoot: string,
    blockTimestamp: bigint,
    blockNumber: bigint
  ];
  export interface OutputObject {
    merkleRoot: string;
    blockTimestamp: bigint;
    blockNumber: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace ContentVerifiedEvent {
  export type InputTuple = [contentHash: BytesLike, exists: boolean];
  export type OutputTuple = [contentHash: string, exists: boolean];
  export interface OutputObject {
    contentHash: string;
    exists: boolean;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface ContentLedgerAnchor extends BaseContract {
  connect(runner?: ContractRunner | null): ContentLedgerAnchor;
  waitForDeployment(): Promise<this>;

  interface: ContentLedgerAnchorInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  admin: TypedContractMethod<[], [string], "view">;

  anchor: TypedContractMethod<[merkleRoot: BytesLike], [bigint], "nonpayable">;

  anchorBlocks: TypedContractMethod<[arg0: BytesLike], [bigint], "view">;

  anchorHistory: TypedContractMethod<[arg0: BigNumberish], [string], "view">;

  anchorTimestamps: TypedContractMethod<[arg0: BytesLike], [bigint], "view">;

  getAnchorCount: TypedContractMethod<[], [bigint], "view">;

  getBlockNumber: TypedContractMethod<
    [merkleRoot: BytesLike],
    [bigint],
    "view"
  >;

  getTimestamp: TypedContractMethod<[merkleRoot: BytesLike], [bigint], "view">;

  isAnchored: TypedContractMethod<[merkleRoot: BytesLike], [boolean], "view">;

  totalAnchors: TypedContractMethod<[], [bigint], "view">;

  transferAdmin: TypedContractMethod<
    [newAdmin: AddressLike],
    [void],
    "nonpayable"
  >;

  verify: TypedContractMethod<[contentHash: BytesLike], [boolean], "view">;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "admin"
  ): TypedContractMethod<[], [string], "view">;
  getFunction(
    nameOrSignature: "anchor"
  ): TypedContractMethod<[merkleRoot: BytesLike], [bigint], "nonpayable">;
  getFunction(
    nameOrSignature: "anchorBlocks"
  ): TypedContractMethod<[arg0: BytesLike], [bigint], "view">;
  getFunction(
    nameOrSignature: "anchorHistory"
  ): TypedContractMethod<[arg0: BigNumberish], [string], "view">;
  getFunction(
    nameOrSignature: "anchorTimestamps"
  ): TypedContractMethod<[arg0: BytesLike], [bigint], "view">;
  getFunction(
    nameOrSignature: "getAnchorCount"
  ): TypedContractMethod<[], [bigint], "view">;
  getFunction(
    nameOrSignature: "getBlockNumber"
  ): TypedContractMethod<[merkleRoot: BytesLike], [bigint], "view">;
  getFunction(
    nameOrSignature: "getTimestamp"
  ): TypedContractMethod<[merkleRoot: BytesLike], [bigint], "view">;
  getFunction(
    nameOrSignature: "isAnchored"
  ): TypedContractMethod<[merkleRoot: BytesLike], [boolean], "view">;
  getFunction(
    nameOrSignature: "totalAnchors"
  ): TypedContractMethod<[], [bigint], "view">;
  getFunction(
    nameOrSignature: "transferAdmin"
  ): TypedContractMethod<[newAdmin: AddressLike], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "verify"
  ): TypedContractMethod<[contentHash: BytesLike], [boolean], "view">;

  getEvent(
    key: "AnchorSubmitted"
  ): TypedContractEvent<
    AnchorSubmittedEvent.InputTuple,
    AnchorSubmittedEvent.OutputTuple,
    AnchorSubmittedEvent.OutputObject
  >;
  getEvent(
    key: "ContentVerified"
  ): TypedContractEvent<
    ContentVerifiedEvent.InputTuple,
    ContentVerifiedEvent.OutputTuple,
    ContentVerifiedEvent.OutputObject
  >;

  filters: {
    "AnchorSubmitted(bytes32,uint256,uint256)": TypedContractEvent<
      AnchorSubmittedEvent.InputTuple,
      AnchorSubmittedEvent.OutputTuple,
      AnchorSubmittedEvent.OutputObject
    >;
    AnchorSubmitted: TypedContractEvent<
      AnchorSubmittedEvent.InputTuple,
      AnchorSubmittedEvent.OutputTuple,
      AnchorSubmittedEvent.OutputObject
    >;

    "ContentVerified(bytes32,bool)": TypedContractEvent<
      ContentVerifiedEvent.InputTuple,
      ContentVerifiedEvent.OutputTuple,
      ContentVerifiedEvent.OutputObject
    >;
    ContentVerified: TypedContractEvent<
      ContentVerifiedEvent.InputTuple,
      ContentVerifiedEvent.OutputTuple,
      ContentVerifiedEvent.OutputObject
    >;
  };
}